orderDetail 구현

+++ src/main/java/com/spring/mapper/CheckoutMapper.java
@@ -6,10 +6,10 @@ import com.spring.domain.CheckoutVO;
 import com.spring.domain.OrderItemVO;
 
 public interface CheckoutMapper {
-	public CheckoutVO read(int orderId);
+	public CheckoutVO read(Long orderId);
 	public OrderItemVO readItem(int pid);
-	public List<OrderItemVO> readOrderList(int orderId);
+	public List<OrderItemVO> readOrderList(Long orderId);
 	public void insert(CheckoutVO checkout);
 	public void itemInsert(OrderItemVO orderItem);
-	public int delete(int orderId);
+	public int delete(Long orderId);
 }
